getPhoto in scriptExecutionControl
