+++ trikScriptRunner/src/scriptEngineWorker.cpp
@@ -85,80 +85,6 @@ QScriptValue timeInterval(QScriptContext *context, QScriptEngine *engine)
 	return engine->toScriptValue(result);
 }
 
-static inline int32_t getMedian(uint8_t &a, uint8_t &b, uint8_t &c, uint8_t &d)
-{
-	if (a > b)
-		std::swap(a, b);
-	if (c > d)
-		std::swap(c, d);
-	if (a > c)
-		std::swap(a, c);
-	if (b > d)
-		std::swap(b, d);
-	return (static_cast<int32_t>(b) + c) >> 1;
-}
-
-QScriptValue getPhoto(QScriptContext *context,	QScriptEngine *engine)
-{
-	const QScriptValue & brickValue = engine->globalObject().property("brick");
-	QObject *qObjBrick = brickValue.toQObject();
-	if (qObjBrick)
-	{
-		auto *brick = qobject_cast<trikControl::BrickInterface*>(qObjBrick);
-		if (brick)
-		{
-			auto port = context->argumentCount() > 0 ? context->argument(0).toString()
-				: QString("/dev/video0");
-			QLOG_INFO() << "Calling getStillImage()";
-			auto data = brick->getStillImage();
-			QList<int32_t> result;
-			result.reserve(data.size() / 3); //Repack RGB88 from 3 x uint8_t into int32_t
-			constexpr auto IMAGE_WIDTH = 320;
-			constexpr auto IMAGE_HEIGHT = 240;
-			if (data.size() >= IMAGE_WIDTH * IMAGE_HEIGHT * 3) {
-				for(int row = 0; row < IMAGE_HEIGHT; row += 2) {
-					for(int col = 0; col < IMAGE_WIDTH; col+=2) {
-						auto row1 = &data[(row*IMAGE_WIDTH+col)*3];
-						auto row2 = row1 + IMAGE_WIDTH*3;
-						auto r1 = row1[0];
-						auto g1 = row1[1];
-						auto b1 = row1[2];
-						auto r2 = row1[3];
-						auto g2 = row1[4];
-						auto b2 = row1[5];
-						auto r3 = row2[0];
-						auto g3 = row2[1];
-						auto b3 = row2[2];
-						auto r4 = row2[3];
-						auto g4 = row2[4];
-						auto b4 = row2[5];
-
-						result.push_back((getMedian(r1, r2, r3, r4) << 16)
-							| (getMedian(g1, g2, g3, g4) << 8)
-							| getMedian(b1, b2, b3, b4));
-					}
-				}
-			}
-
-			QLOG_INFO() << "Constructed result of getStillImage()";
-			auto val = engine->toScriptValue(result);
-			QLOG_INFO() << "Result of getStillImage() converted to JS value";
-			return val;
-		}
-		else
-		{
-			QLOG_ERROR() << "script getPhoto failed at downcasting qObject to Brick";
-			return QScriptValue();
-		}
-	}
-	else
-	{
-		QLOG_ERROR() << "script getPhoto failed to get brick Obj";
-		return QScriptValue();
-	}
-
-}
-
 ScriptEngineWorker::ScriptEngineWorker(trikControl::BrickInterface &brick
 		, trikNetwork::MailboxInterface * const mailbox
 		, ScriptExecutionControl &scriptControl
@@ -178,7 +104,6 @@ ScriptEngineWorker::ScriptEngineWorker(trikControl::BrickInterface &brick
 
 	registerUserFunction("print", print);
 	registerUserFunction("timeInterval", timeInterval);
-	registerUserFunction("getPhoto", getPhoto);
 }
 
 void ScriptEngineWorker::brickBeep()
