+++ trikScriptRunner/src/scriptExecutionControl.h
@@ -17,6 +17,7 @@
 #include <QtCore/QList>
 #include <QtCore/QTimer>
 #include <QtCore/QStringList>
+#include <trikControl/brickInterface.h>
 
 namespace trikScriptRunner {
 
@@ -26,11 +27,14 @@ class ScriptExecutionControl : public QObject
 	Q_OBJECT
 
 public:
+	ScriptExecutionControl(trikControl::BrickInterface &brick);
 	~ScriptExecutionControl() override;
 
 	/// Returns true if a script is in event-driven running mode, so it shall wait for events when script is executed.
 	/// If it is false, script will exit immediately.
 	bool isInEventDrivenMode() const;
+
+	QList<int32_t> getPhoto();
 
 	/// Starts a new timer with given interval and returns reference to it.
 	Q_INVOKABLE QTimer *timer(int milliseconds);
@@ -81,6 +85,8 @@ signals:
 
 private:
 	QList<QTimer *> mTimers; // Has ownership.
+	trikControl::BrickInterface &mBrick;
+
 
 	/// True, if a system is in event-driven running mode, so it shall wait for events when script is executed.
 	/// If it is false, script will exit immediately.
